+++ io/module.f.ts
@@ -10,6 +10,7 @@ import { fromVec, toVec } from '../types/uint8array/module.f.ts'
  */
 export type Dirent = {
     readonly name: string
+    readonly parentPath: string
     readonly isDirectory: () => boolean
     readonly isFile: () => boolean
 }
@@ -148,6 +149,9 @@ asyncRun({
     log: async message => log(message),
     mkdir: param => tc(async() => { await mkdir(...param) }),
     readFile: path => tc(async() => toVec(await readFile(path))),
-    readdir: param => tc(async() => (await readdir(...param)).map(normalize)),
+    readdir: ([path, r]) => tc(async() =>
+        (await readdir(path, { ...r, withFileTypes: true }))
+        .map(v => ({ name: v.name, parentPath: normalize(v.parentPath), isFile: v.isFile() }))
+    ),
     writeFile: ([path, data]) => tc(() => writeFile(path, fromVec(data))),
 })