+++ types/effect/node/module.f.ts
@@ -26,16 +26,28 @@ export const readFile =
 
 // readdir
 
-export type ReaddirOptions = { readonly recursive?: true }
+/**
+ * Represents a directory entry (file or directory) in the filesystem
+ * @see https://nodejs.org/api/fs.html#class-fsdirent
+ */
+export type Dirent = {
+    readonly name: string
+    readonly parentPath: string
+    readonly isFile: boolean
+}
+
+export type ReaddirOptions = {
+    readonly recursive?: true
+}
 
 export type ReaddirParam = readonly[string, ReaddirOptions]
 
 export type Readdir = {
-    readonly readdir: readonly[ReaddirParam, IoResult<readonly string[]>]
+    readonly readdir: readonly[ReaddirParam, IoResult<readonly Dirent[]>]
 }
 
 export const readdir =
-    (...p: ReaddirParam): Do<Readdir, IoResult<readonly string[]>> =>
+    (...p: ReaddirParam): Do<Readdir, IoResult<readonly Dirent[]>> =>
     do_('readdir', p)
 
 // writeFile